[WikiBundle] fixes section delete and normalize pdf export
wiki cs
