+++ main/app/Resources/translations/actions.en.json
@@ -30,6 +30,7 @@
   "exit": "Exit",
   "expand": "Expand",
   "export": "Export",
+  "export-pdf": "Export in PDF",
   "filter": "Filter",
   "finish": "Finish",
   "follow": "Follow",
