+++ main/app/Resources/translations/actions.fr.json
@@ -30,6 +30,7 @@
   "exit": "Quitter",
   "expand": "Déplier",
   "export": "Exporter",
+  "export-pdf": "Exporter en PDF",
   "filter": "Filtrer",
   "finish": "Terminer",
   "follow": "Suivre",
