+++ main/core/Resources/translations/platform.de.json
@@ -12,7 +12,6 @@
     "activate_account": "Aktivieren Sie Ihr Benutzerkonto",
     "activate_account_click": "Klicken Sie hier, um Ihr Benutzerkonto zu aktivieren",
     "activate_notifications": "Benachrichtigungen aktivieren",
-    "activate_pdf_export": "PDF Export aktivieren",
     "activate_personal_workspace_creation": "",
     "active_widgets": "Aktive Widgets",
     "activities": "Aktivitäten",
@@ -589,7 +588,6 @@
     "password_ok": "",
     "paste": "",
     "path": "",
-    "pdf_export": "",
     "pending": "",
     "pending_friends": "",
     "pending_users_list": "",
