+++ main/core/Resources/translations/platform.en.json
@@ -14,7 +14,6 @@
     "activate_account_click": "Click here to activate your account",
     "activate_comments": "Activate comments",
     "activate_notifications": "Activate notifications",
-    "activate_pdf_export": "Activate PDF export",
     "activate_personal_workspace_creation": "Activate personal workspace creation.",
     "activated": "Activated",
     "activities": "Activities",
@@ -771,7 +770,6 @@
     "paste": "Paste",
     "path": "Path",
     "paths_tracking": "Paths tracking",
-    "pdf_export": "Export in PDF",
     "pending": "Pending",
     "pending_friends": "Pending friends",
     "pending_users_list": "Pending users list",
