+++ main/core/Resources/translations/platform.es.json
@@ -12,7 +12,6 @@
     "activate_account": "Activación de la cuenta",
     "activate_account_click": "",
     "activate_notifications": "Activar notificaciones",
-    "activate_pdf_export": "Activar exportación a PDF",
     "activate_personal_workspace_creation": "Activar la creación de espacios personales.",
     "active_widgets": "Widgets activos",
     "activities": "Actividades",
@@ -595,7 +594,6 @@
     "password_ok": "Su contraseña ha cambiado, ahora pude conectarse",
     "paste": "Pegar",
     "path": "Ruta",
-    "pdf_export": "Exportación a PDF",
     "pending": "",
     "pending_friends": "",
     "pending_users_list": "",
