+++ main/core/Resources/translations/platform.fr.json
@@ -14,7 +14,6 @@
     "activate_account_click": "Cliquez ici pour activer votre compte.",
     "activate_comments": "Activer les commentaires",
     "activate_notifications": "Activer les notifications",
-    "activate_pdf_export": "Activer l'export PDF",
     "activate_personal_workspace_creation": "Activer la création des espaces personnels.",
     "activated": "Activé",
     "activities": "Activités",
@@ -789,7 +788,6 @@
     "paste": "Coller",
     "path": "Parcours",
     "paths_tracking": "Suivi des parcours",
-    "pdf_export": "Export PDF",
     "pending": "En attente",
     "pending_friends": "Amis en attente",
     "pending_users_list": "Liste des utilisateurs en attente",
