+++ main/core/Resources/translations/platform.nl.json
@@ -13,7 +13,6 @@
     "activate_account": "Activeer uw account",
     "activate_account_click": "Klik hier om uw account te activeren",
     "activate_notifications": "Notities activeren",
-    "activate_pdf_export": "PDF-uitvoer activeren",
     "activate_personal_workspace_creation": "Het maken van persoonlijke werkruimten activeren.",
     "activated": "Geactiveerd",
     "active_widgets": "Actieve widgets",
@@ -698,7 +697,6 @@
     "password_ok": "Uw wachtwoord is gewijzigd. U kunt zich nu inloggen.",
     "paste": "Coller",
     "path": "Plaats",
-    "pdf_export": "PDF uitvoer",
     "pending": "In wacht",
     "pending_friends": "Vrienden in wacht",
     "pending_users_list": "Lijst van de gebruikers in wacht",
