+++ plugin/wiki/Resources/config/routing.yml
@@ -1,11 +1,6 @@
-# Resource
-icap_wiki_resource:
-    resource: "@IcapWikiBundle/Controller/Resource/WikiController.php"
-    type:     annotation
-
 # JSON API
 icap_wiki_api:
-    resource: "@IcapWikiBundle/Controller/API/"
+    resource: "@IcapWikiBundle/Controller/"
     prefix: apiv2
     type: api
     options:
