+++ plugin/wiki/Resources/modules/resources/wiki/store/actions.js
@@ -4,7 +4,7 @@ export const actions = {}
 
 actions.downloadWikiPdf = (wikiId) => ({
   [API_REQUEST]: {
-    url: ['icap_wiki_export_pdf', {id: wikiId}],
+    url: ['apiv2_wiki_export_pdf', {id: wikiId}],
     request: {
       method: 'GET'
     }
