+++ plugin/wiki/Resources/styles/resources/wiki/index.less
@@ -78,15 +78,10 @@
 // ---
 
 .wiki-section {
+    margin-bottom: @component-margin-bottom;
     &.wiki-section-invisible {
         .wiki-section-title {
             background-color: @state-danger-bg;
-            .wiki-section-invisible-text {
-                /*display: inline-block;
-                font-size: 0.7em;
-                font-style: italic;
-                color: @gray-light;*/
-            }
         }
     }
 }
